refactor(*): Make blocks pointed to by a pointer instead of an index
Functions and blocks used to be identified by their index in the bytes array. This complicates replacing functions with code from malloced places.

There is still an issue with the block_lookup field of Module. It maps indexes to blocks and this will still clash with our goal.

The cur_pc has be renamed block_ptr in i_instr funcions that used it. As cur_pc acctualy pointed to the current first byte of the executeb block (that is m->pc_ptr at the start of the while-swich loop).