add remote function call class
RFC runs either on the emulator or MCU. When executed on the emulator it takes a monitoring rol i.e., it keeps track of function identifiers that need to be executed remotely on the MCU. Functions that need to be executed remote are registered via `RFC::registerRFCs` or `RFC::registerRFC`.
When executed on the MCU, the class takes the rol of executing a function on the MCU: functions that need to be executed on the MCU are registered via `RFC::registerRFCallee`.