Enalbe the update of the proxy list more than once
The `interruptMonitorProxies` can now be called more than once for the same MCU host and this solely to update the proxy list. No new connection to the host is triggered. Triggering the same interrupt with a different host and/or port will cause the emulator to first close the existing socket connections.