Add unit testing in cmakelist
Now we can build the project to run the unit tests defined in directory `tests/vm_implementation_unit_tests/`.
For each file in that directory an executable is generated and register as test in `ctest`.
After building all testing executables, all executable can be run through `ctest` command.

If we need to create a testing executable based on multiple source code files.
We should then update the `CMakeLists.txt` file.