bug fix: removing ISR from pins after module update
Via 2 new methods `setUnsubscribe`  and `unsubscribe`, we can respectively set a lambda that unsubsribes an ISR of a pin and call that lambda.
We use this ability to reimplement clear_backs which is now also called after updating a module.
In Arduino.cpp we register `unsubscribe` callbacks that run `detachinterrupt` on a pin.