Added temporary hardcoded motor speed for motor 0 and 1 when stepping back.
In the future we should replace this with the motor speed used in the arguments of the function call. This requires taking snapshots before every primitive call instead of after and handling the last snapshot in a special way by not restoring IO and using the overrides from a future snapshot to make sure that the multiverse is traversed correctly.