Reversible primitives (#250)
* Initial work on io state inspection

This allows the debugger to inspect the current io state and load io states from snapshots. Currently only works with chip_digital_write.

* Obtain io state through portOutput/InputRegister

There currently is no support for analog values.

* Remove io map from Module struct

* Use NUM_DIGITAL_PINS because SOC_GPIO_PIN_COUNT is esp specific

* Serialize/deserialize pin mode in snapshot data

* Use digitalPinToBitMask in an attempt to make the code more microcontroller agnostic

* Port WARDuino to the Zephyr RTOS

The main reason for this is to support boards such as Open Bot Brain which uses
an stm32l496zg. By porting to Zephyr we can also add support for some other
boards with relative ease.

* Try correcting for drift in the servo motor driver

* Fixed cmake with branch rebase

* Use vfprintf instead of fprintf to handle variadic arguments

* Loading snapshots should work again now

* Hacky but it's starting to work

* Somewhat hacky way of extracting the current sensor value

* Put uart heartbeat for sensors in a thread + colour_sensor now returns sensor value

* Allow controlling multiple motors

* Added anti-stall system to motor driver

This makes it so if the motor tries to move but it's not actually moving the
voltage applied to the motor is increased until it moves and completes the
operation.

* Make motor driver reversible

* Added some things needed to make uart sensor work

* Added invoke_primitive function that makes it easy to call primitives with arguments

* New API for reversible primitives

Reversible primitives are now defined by using def_prim_reverse and def_prim_serialize.

* Rename PinState to IOStateElement and clean up debugger ioState interrupt

* Clang format

* Only run the reverse operation for the primitive that was just executed and not for all primitives

This particularly important for motor primitives that for example drive two motors at the same time vs a primitive that drives one motor at a time, executing the wrong reverse operation will result in the correct angles but the wrong position.

* Use ms to configure stall timeout for motors

* Adjust speed faster when stalling

Also added helper functions to reduce duplicated code between drive_motor_degrees and drive_motor_degrees_relative.

* Don't wait for the full 10ms if we have already detected movement

* Stripped out open bot brain primitives

* Clang format

* Make chip_digital_write on arduino reversible using the new macros

* Remove io-arduino.cpp and fix usage of PinState

* def_prim_serialize should not return anything for chip_digital_write

* Add extra clause to if so int32_t and int are bot considered I32 in create_stack template function

* Add new macros + restore_external_state + get_io_state to ESP-IDF

* Install primitive reverse for chip_digital_write

* Just use NUM_PRIMITIVES and set it to 5

* Move last remaining bits of io.h into primitives.h

* Add reversible primitive code to emulated.cpp

* Fixed zephyr build