🚀 Add checkpointing policy (#253)
✨ adds checkpointing policy (for multiverse debugging)
✨ adds the "continue for" debug operation

**commits:**

* Add basic checkpointing snapshotting policy

A checkpoint is taken at primitives and every x instructions. When the VM is paused it also makes a checkpoint so the debugger knows where the other snapshots are relative to the current point in time and so it can also display what the current state of the vm is.

TODO: Will need to be extended a bit when using multiverse debugging. Choicepoints should have checkpoints right after reading sensor values so you don't have to remember the sensor value.

* Remove unused duplicate interpret function

* Add instructions_executed into a json object

* Checkpoint on step so the debugger knows the current state again + checkpoint after instead of before primitive calls

* Added continue for debug interrupt that allows executing x instructions of the program

* Take a checkpoint upon enabling checkpointing mode

* Fix uninitialized variable

* No checkpointing during WARDUINOinit

* Only send STEP! after actually completing a step and not before

Also send any checkpoints right before the STEP!.

* Allow setting the checkpoint interval instead of hardcoding it to 10

* Make checkpoint interval a 32bit integer

* Send arguments + primitive index that was called in checkpoint messages

* Got rid of unneeded use of prev_pc_ptr

* Clang-format

* Use PRIu32 to display amount of instructions to step in continue for debug message

* Rename isPrimitiveBeingCalled and remove unnecessary Module argument

* Minor cleanup

* Clang-format

* Use raw string literal

* Fix checkpointing so checkpoints are only taken after primitive calls