Add some hidpi scaling support
Works but it resets the font size every time draw_text is called which is not good for performance because it resets the font cache. The better solution is probably to use SDL3 which as an event fot the dpi changing which allows us to change the font size at that point. This is much more efficient than constantly updating it.