Open Bot Brain touch sensor support. (#321)
* Initial work on getting analog read working

This will also make the touch sensors work!

* work in progress adc3 configuration for input 1

* Switch between adcs for the different ports

Currently done with an if, we probably want to do this with an array in the end.

* Add pin 6 channels for ev3 touch sensors

* EV3 Touch sensor support

* Rename analog_read to nxt_touch_sensor

Will add support for generic analog_read later.

* Rename analog_read to nxt_touch_sensor

Will add support for generic analog_read later.

* clang-format

* Minor changes + add print_string to zephyr

* Added nxt and ev3 touch sensor primitives to the emulator

Instead of always returning 1 or 0 they also use a random 12bit number.